+++ src/trx/game/enum.c
@@ -176,6 +176,8 @@ static __attribute__((constructor)) void M_Init(void)
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_TICK, "tick");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_KEY_DOWN, "key_down");
     ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_KEY_UP, "key_up");
+    ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_BUTTON_DOWN, "button_down");
+    ENUM_MAP(LUA_EVENT_TYPE, LUA_EVENT_BUTTON_UP, "button_up");
 
     ENUM_MAP(UI_REGION, UI_REGION_TOP_LEFT, "top-left");
     ENUM_MAP(UI_REGION, UI_REGION_TOP_CENTER, "top-center");